requirements.txt
asgiref==3.4.1
Deprecated==1.2.13
Django==3.2.9
django-cors-headers==3.10.1
django-redis==5.2.0
django-smtp-ssl==1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
mysqlclient==2.1.0
packaging==21.3
Pillow==8.4.0
PyJWT==2.1.0
pyparsing==3.0.6
pytz==2021.3
redis==4.1.0
sqlparse==0.4.2
tzdata==2021.5
wrapt==1.13.3
